+++ Classes/ForksController.h
@@ -3,8 +3,5 @@
 
 
 @interface ForksController : UITableViewController
-@property(nonatomic,strong)IBOutlet UITableViewCell *loadingForksCell;
-@property(nonatomic,strong)IBOutlet UITableViewCell *noForksCell;	
-
 - (id)initWithRepository:(GHRepository *)repo;
 @end